requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.14
Faker==40.1.2
inflection==0.5.1
Markdown==3.10.1
packaging==26.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
validate_docbr==1.11.1
